# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payments Gate",
        "description": "API for handling payment transactions",
        "version": "1.0.0"
    },
    "paths": {
        "/api/transactions/recent": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get recent transactions",
                "operationId": "94c31977814f04f492128758c14bef0e",
                "parameters": [
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "Merchant API key",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "api-key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recent transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "transactionUuid": {
                                                "type": "string",
                                                "example": "c061745d-e26f-4d8b-b923-4f2297f414a6"
                                            },
                                            "amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 199.99
                                            },
                                            "currency": {
                                                "type": "string",
                                                "example": "PLN"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "SUCCESS"
                                            },
                                            "paymentMethod": {
                                                "type": "string",
                                                "example": "TPAY"
                                            },
                                            "createdAt": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-03-01 12:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/payment-methods/share": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get payment method share statistics",
                "operationId": "4431ba811abc4b95a722a451ddcd6897",
                "parameters": [
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "Merchant API key",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "api-key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Share of each payment method",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "paymentMethod": {
                                                "type": "string",
                                                "example": "TPAY"
                                            },
                                            "count": {
                                                "type": "integer",
                                                "example": 100
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/total": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get total number of transactions",
                "operationId": "6f72125fdb1a4f8ab192da0eb981064f",
                "parameters": [
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "Merchant API key",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "api-key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Total transaction count",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total": {
                                            "type": "integer",
                                            "example": 1234
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/balances": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get total balances of transactions",
                "operationId": "59a8b4cc0c0b9b0fc2dc43b3a11b307e",
                "parameters": [
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "Merchant API key",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "api-key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Balances of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "totalAmount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 9999.99
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "PLN"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/rejected": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get total number of failed transactions",
                "operationId": "f10655a93dc7e2a6ac48ae54497f0c97",
                "parameters": [
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "Merchant API key",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "api-key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Total failed transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "failedCount": {
                                            "type": "integer",
                                            "example": 12
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get all transactions for merchant",
                "operationId": "79d70cb12da898f44d1c95f18549b0ed",
                "parameters": [
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "Merchant API key",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "api-key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of merchant transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "transactionUuid": {
                                                "type": "string",
                                                "example": "c061745d-e26f-4d8b-b923-4f2297f414a6"
                                            },
                                            "amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 199.99
                                            },
                                            "currency": {
                                                "type": "string",
                                                "example": "PLN"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "SUCCESS"
                                            },
                                            "paymentMethod": {
                                                "type": "string",
                                                "example": "TPAY"
                                            },
                                            "createdAt": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-03-01 12:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transaction notifications for merchant",
                "operationId": "303534d329252cfb951a36f27a9b3bbd",
                "parameters": [
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "Merchant API key",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "api-key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transaction notifications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "transactionUuid": {
                                                "type": "string",
                                                "example": "c6e2c816-3f5e-417b-9e91-a794223aa903"
                                            },
                                            "transactionId": {
                                                "type": "integer",
                                                "example": 123
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "SUCCESS"
                                            },
                                            "statusType": {
                                                "type": "string",
                                                "example": "SUCCESS_REFUND"
                                            },
                                            "createdAt": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-03-01 12:00:00",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Wprowadź swój klucz API",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "api_key": []
        }
    ]
}